requirements/common.txt
# This is an implicit value, here for clarity
--index-url https://pypi.python.org/simple/

# Bleeding edge Django
Django>=1.7,<1.8
django-extensions>=1.4

# Configuration
django-environ>=0.3

# Forms
django-braces>=1.4

# Models
django-model-utils>=2.2

psycopg2
simplejson
djangorestframework
csvkit
sqlparse
numpy

pandas==0.14.1

ipython[notebook]

git+https://github.com/openpolis/django-popolo.git@develop#egg=django-popolo